@@ -1,4 +1,4 @@
-ruby "2.0.0"
+ruby "2.1.5"
 
 source "https://rubygems.org"
 source "https://rails-assets.org"
