@@ -22,17 +22,46 @@ module Sprangular
         end
       ]
 
-      {env: Rails.env,
+      {
+        env: Rails.env,
         config: {
           site_name: store.seo_title || store.name,
-          logo:      asset_path(config.logo),
+          logo: asset_path(config.logo),
+          locale: I18n.locale,
+          currency: Money::Currency.table[current_currency.downcase.to_sym],
+          supported_locales: supported_locales,
           default_country_id: config.default_country_id,
-          facebook_app_id: ENV['FACEBOOK_APP_ID'],
           payment_methods: payment_methods,
-          image_sizes: Spree::Image.attachment_definitions[:attachment][:styles].keys
+          image_sizes:
+            Spree::Image.attachment_definitions[:attachment][:styles].keys,
+          product_page_size: Spree::Config.products_per_page
         },
+        translations: current_translations,
         templates: templates
       }
+    end
+
+    def supported_locales
+      if defined? SpreeI18n
+        SpreeI18n::Config.supported_locales
+      else
+        Rails.application.config.i18n.available_locales
+      end
+    end
+
+    ##
+    # Get relevant translations for front end. For both a simple, and
+    # "Chainable" i18n Backend, which is used by spree i18n.
+    def current_translations
+      if I18n.backend.class == I18n::Backend::Simple
+        I18n.backend.load_translations
+
+        @translations ||= I18n.backend.send(:translations)
+      else
+        @translations ||= I18n.backend.backends.last.send(:translations)
+      end
+      # Return only sprangular keys for js environment
+      @translations[I18n.locale][:sprangular]
     end
 
     def cached_templates
@@ -46,7 +75,7 @@ module Sprangular
 
       files = Dir[root + "app/assets/templates/#{dir}/**"].inject(files) do |hash, path|
         asset_path = asset_path path.gsub(root.to_s + "/app/assets/templates/", "")
-        local_path = 'app/assets/templates/' + asset_path
+        local_path = "app/assets/templates/" + asset_path
 
         hash[asset_path.gsub(/.slim$/, '')] = Tilt.new(path).render.html_safe if !File.exists?(local_path)
 
@@ -54,10 +83,13 @@ module Sprangular
       end
 
       Dir["app/assets/templates/#{dir}/**"].inject(files) do |hash, path|
-        asset_path = asset_path(path.gsub("/app/assets/templates/", ""))
-        asset_path = asset_path.gsub(/^\/app\/assets\/templates/, '/assets').gsub(/.slim$/, '')
+        sprockets_path = path.gsub("app/assets/templates/", "")
 
-        hash[asset_path] = Tilt.new(path).render.html_safe
+        asset_path = asset_path(sprockets_path).
+          gsub(/^\/app\/assets\/templates/, '/assets').
+          gsub(/.slim$/, '')
+
+        hash[asset_path] = Rails.application.assets.find_asset(sprockets_path).body.html_safe
         hash
       end
     end
