@@ -1,6 +1,9 @@
 class Sprangular::InstallGenerator < Rails::Generators::Base
   source_root File.expand_path('../templates', __FILE__)
 
+  argument :module_name, type: :string, default: File.basename(Dir.pwd).underscore.camelize
+  class_option :copy_templates, type: :boolean, default: false
+
   def remove_spree_umbrella_gem
     gsub_file 'Gemfile', /^.*gem.["']spree["'].*\n/, ''
   end
@@ -15,9 +18,9 @@ class Sprangular::InstallGenerator < Rails::Generators::Base
   end
 
   def remove_assets
-    git rm: "-rf vendor/assets/images/spree/frontend"
-    git rm: "-rf vendor/assets/javascripts/spree/frontend"
-    git rm: "-rf vendor/assets/stylesheets/spree/frontend"
+    run "rm -rf vendor/assets/images/spree/frontend"
+    run "rm -rf vendor/assets/javascripts/spree/frontend"
+    run "rm -rf vendor/assets/stylesheets/spree/frontend"
   end
 
   def remove_uneeded_js
@@ -33,12 +36,30 @@ class Sprangular::InstallGenerator < Rails::Generators::Base
 //= require jquery
 //= require bootstrap-sass-official
 //= require sprangular
+//= require sprangular/host
 //= require sprangular/extraRoutes
+//= require_tree ./sprangular/controllers
+//= require_tree ./sprangular/services
 eos
 
-    copy_file 'extraRoutes.coffee', 'app/assets/javascripts/sprangular/extraRoutes.coffee'
-    copy_file 'about.html.slim', 'app/assets/templates/static/about.html.slim'
-    copy_file 'terms.html.slim', 'app/assets/templates/static/terms.html.slim'
-    copy_file 'privacy.html.slim', 'app/assets/templates/static/privacy.html.slim'
+    template 'host.coffee',        'app/assets/javascripts/sprangular/host.coffee'
+    template 'extraRoutes.coffee', 'app/assets/javascripts/sprangular/extraRoutes.coffee'
+
+    create_file 'app/assets/javascripts/sprangular/controllers/.keep', ''
+    create_file 'app/assets/javascripts/sprangular/services/.keep', ''
+
+    copy_file 'about.html.slim',    'app/assets/templates/static/about.html.slim'
+    copy_file 'terms.html.slim',    'app/assets/templates/static/terms.html.slim'
+    copy_file 'privacy.html.slim',  'app/assets/templates/static/privacy.html.slim'
+  end
+
+  def add_layout
+    template 'layout.slim', 'app/views/layouts/sprangular/application.slim'
+  end
+
+  def copy_templates
+    if options["copy_templates"]
+      directory "../../../../../app/assets/templates", "app/assets/templates"
+    end
   end
 end
