@@ -6,9 +6,18 @@ require 'geocoder'
 require 'slim'
 require 'rabl'
 require 'jquery-rails'
-require 'angularjs-rails'
 require 'rails-assets-underscore'
 require 'rails-assets-underscore.string'
+require 'ngannotate/rails'
+require 'prerender_rails'
+require 'rails-assets-angular'
+require 'rails-assets-angular-route'
+require 'rails-assets-angular-sanitize'
+require 'rails-assets-angular-animate'
+require 'rails-assets-underscore'
+require 'rails-assets-underscore.string'
+require 'rails-assets-angular-translate'
+require 'rails-assets-angular-i18n'
 require 'rails-assets-angularytics'
 require 'bourbon'
 require 'neat'
