@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe "Home Page", js: true do
-  it "shows welcome message" do
+  scenario "shows welcome message" do
     visit sprangular_engine.root_path
     expect(page).to have_content("Welcome to Sprangular!")
   end
